requirements.txt
asgiref==3.5.2
braintree==4.16.0
certifi==2020.12.5
charset-normalizer==2.1.0
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
idna==3.3
Pillow==9.2.0
pycodestyle==2.7.0
pytz==2020.5
requests==2.28.1
sqlparse==0.4.1
toml==0.10.2
tzdata==2022.2
urllib3==1.26.4
